+++ Model/ExtractInParameter.cs
@@ -5,7 +5,7 @@ namespace recipeconfigurationservice.Model
     public class ExtractInParameter
     {
         [Key]
-        public long extractInParameterId {get;set;}
+        public int extractInParameterId {get;set;}
         public string path{get;set;}
         [Required]
         public string type{get;set;}
