+++ Model/ExtractOutParameter.cs
@@ -5,7 +5,7 @@ namespace recipeconfigurationservice.Model
     public class ExtractOutParameter
     {
         [Key]
-        public long extractOutParameterId {get;set;}
+        public int extractOutParameterId {get;set;}
         public string path{get;set;}
         [Required]
         public string name{get;set;}
