+++ Services/Interfaces/IExtractService.cs
@@ -7,12 +7,12 @@ namespace recipeconfigurationservice.Services.Interfaces
     public interface IExtractService
     {
         Task<List<Extract>> getExtracts(int start,int quantity);
-        Task<Extract> getExtract(long extractId);
+        Task<Extract> getExtract(int extractId);
 
          Task<Extract> addExtract(Extract extract);
 
-         Task<Extract> updateExtract(long extractId,Extract extract);
-         Task<Extract> deleteExtract(long extractId);
+         Task<Extract> updateExtract(int extractId,Extract extract);
+         Task<Extract> deleteExtract(int extractId);
 
         
     }
